+++ service/src/llm/llm.controller.ts
@@ -4,14 +4,14 @@ import { RequestData } from "../common/index.dto.js";
 import { PostMessageble } from "../hook/adapter.js";
 import { getClient } from "../mcp/connect.service.js";
 import { abortMessageService, streamingChatCompletion } from "./llm.service.js";
-
+import { OpenAI } from "openai";
 export class LlmController {
 
     @Controller('llm/chat/completions')
     async chatCompletion(data: RequestData, webview: PostMessageble) {
 
         try {
-            await streamingChatCompletion(data, webview);            
+            await streamingChatCompletion(data, webview);
         } catch (error) {
             console.log('error' + error);
             
@@ -35,4 +35,20 @@ export class LlmController {
         return abortMessageService(data, webview);
     }
 
+
+    @Controller('llm/models')
+    async getModels(data: RequestData, webview: PostMessageble) {
+        const {
+            baseURL,
+            apiKey,
+        } = data;        
+        
+        const client = new OpenAI({ apiKey, baseURL });
+        const models = await client.models.list();
+
+        return {
+            code: 200,
+            msg: models.data
+        }
+    }
 }