+++ service/src/llm/llm.service.ts
@@ -5,6 +5,7 @@ import { RestfulResponse } from "../common/index.dto.js";
 import { ocrDB } from "../hook/db.js";
 import type { ToolCallContent } from "../mcp/client.dto.js";
 import { ocrWorkerStorage } from "../mcp/ocr.service.js";
+import { axiosFetch } from "../hook/axios-fetch.js";
 
 export let currentStream: AsyncIterable<any> | null = null;
 
@@ -12,33 +13,51 @@ export async function streamingChatCompletion(
     data: any,
     webview: PostMessageble
 ) {
-    let {
-            baseURL,
-            apiKey,
-            model,
-            messages,
-            temperature,
-            tools = [],
-            parallelToolCalls = true
-        } = data;
+    const {
+        baseURL,
+        apiKey,
+        model,
+        messages,
+        temperature,
+        tools = [],
+        parallelToolCalls = true,
+        proxyServer = ''
+    } = data;
 
     const client = new OpenAI({
         baseURL,
-        apiKey
+        apiKey,
+        fetch: async (input: string | URL | Request, init?: RequestInit) => {
+
+            console.log('openai fetch begin, proxyServer:', proxyServer);
+            
+            if (model.startsWith('gemini') && init) {
+                // 该死的 google
+                init.headers = {
+                    'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${apiKey}`
+                }
+            }
+            
+            return await axiosFetch(input, init, { proxyServer });
+        }
     });
 
-    if (tools.length === 0) {
-        tools = undefined;
-    }
-    
+    const seriableTools = (tools.length === 0) ? undefined: tools;
+    const seriableParallelToolCalls = (tools.length === 0)? 
+        undefined: model.startsWith('gemini') ? undefined : parallelToolCalls;
+     
     await postProcessMessages(messages);
+
+    console.log('seriableTools', seriableTools);
+    console.log('seriableParallelToolCalls', seriableParallelToolCalls);
     
     const stream = await client.chat.completions.create({
         model,
         messages,
         temperature,
-        tools,
-        parallel_tool_calls: parallelToolCalls,
+        tools: seriableTools,
+        parallel_tool_calls: seriableParallelToolCalls,
         stream: true
     });